+++ alacritty/src/event.rs
@@ -352,7 +352,7 @@ impl<'a, N: Notify + 'a, T: EventListener> input::ActionContext<T> for ActionCon
             args.push(arg.into());
         }
 
-        start_daemon(&alacritty, &args);
+        start_daemon(&alacritty, &args, None);
     }
 
     /// Spawn URL launcher when clicking on URLs.
@@ -367,7 +367,7 @@ impl<'a, N: Notify + 'a, T: EventListener> input::ActionContext<T> for ActionCon
             let end = self.terminal.visible_to_buffer(url.end());
             args.push(self.terminal.bounds_to_string(start, end));
 
-            start_daemon(launcher.program(), &args);
+            start_daemon(launcher.program(), &args, None);
         }
     }
 
@@ -551,6 +551,14 @@ impl<'a, N: Notify + 'a, T: EventListener> input::ActionContext<T> for ActionCon
 
     fn scheduler_mut(&mut self) -> &mut Scheduler {
         self.scheduler
+    }
+
+    fn to_string(&self) -> String {
+        self.terminal.grid_to_string()
+    }
+
+    fn to_string_only_visible(&self) -> String {
+        self.terminal.grid_to_string_only_visible()
     }
 }
 
@@ -1031,7 +1039,8 @@ impl<N: Notify + OnResize> Processor<N> {
                     TerminalEvent::Wakeup => processor.ctx.terminal.dirty = true,
                     TerminalEvent::Bell => {
                         let bell_command = processor.ctx.config.bell().command.as_ref();
-                        let _ = bell_command.map(|cmd| start_daemon(cmd.program(), cmd.args()));
+                        let _ =
+                            bell_command.map(|cmd| start_daemon(cmd.program(), cmd.args(), None));
                         if processor.ctx.terminal.mode().contains(TermMode::URGENCY_HINTS) {
                             processor.ctx.window.set_urgent(!processor.ctx.terminal.is_focused);
                         }