+++ alacritty_terminal/src/term/mod.rs
@@ -1457,6 +1457,43 @@ impl<T> Term<T> {
             rendered: false,
         }
     }
+
+    fn grid_to_string_between(&self, y0: Line, y1: Line) -> String {
+        let grid = &self.grid;
+
+        let mut s = String::with_capacity(*(y1 - y0) * *grid.cols());
+
+        for y in (*y0..*y1).rev() {
+            let row = &grid[y];
+
+            // Add the chars to the string.
+            for x in 0..*row.line_length() {
+                s.push(row[Column(x)].c);
+            }
+
+            if !row.last().unwrap().flags.contains(Flags::WRAPLINE) {
+                s.push('\n');
+            }
+        }
+
+        // Ensure that the string ends with a newline.
+        if !s.ends_with('\n') {
+            s.push('\n');
+        }
+
+        s
+    }
+
+    pub fn grid_to_string_only_visible(&self) -> String {
+        self.grid_to_string_between(Line(self.grid.display_offset()), self.grid.screen_lines())
+    }
+
+    pub fn grid_to_string(&self) -> String {
+        self.grid_to_string_between(
+            Line(0),
+            Line(*self.grid.screen_lines() + self.grid.history_size()),
+        )
+    }
 }
 
 impl<T> Dimensions for Term<T> {
@@ -2823,6 +2860,43 @@ mod tests {
         term.title = Some("Test".into());
         term.set_title(None);
         assert_eq!(term.title, None);
+    }
+
+    #[test]
+    fn to_string() {
+        let mut grid = Grid::<Cell>::new(Line(4), Column(4), 1);
+        grid[Line(0)][Column(0)].c = '1';
+        grid[Line(0)][Column(1)].c = ' ';
+        grid[Line(0)][Column(2)].c = '2';
+
+        grid[Line(1)][Column(0)].c = '3';
+        grid[Line(1)][Column(1)].c = '4';
+        grid[Line(1)][Column(2)].c = '5';
+        grid[Line(1)][Column(3)].c = '6';
+        grid[Line(1)][Column(3)].flags.insert(Flags::WRAPLINE);
+
+        grid[Line(2)][Column(0)].c = 'a';
+        grid[Line(2)][Column(1)].c = 'b';
+        grid[Line(2)][Column(2)].c = 'c';
+        grid[Line(2)][Column(3)].c = 'd';
+
+        grid[Line(3)][Column(0)].c = 'e';
+        grid[Line(3)][Column(1)].c = 'f';
+        grid[Line(3)][Column(2)].c = 'g';
+        grid[Line(3)][Column(3)].c = 'h';
+        grid[Line(3)][Column(3)].flags.insert(Flags::WRAPLINE);
+
+        let size = SizeInfo::new(21.0, 51.0, 3.0, 3.0, 0.0, 0.0, false);
+        let mut term = Term::new(&MockConfig::default(), size, Mock);
+        mem::swap(&mut term.grid, &mut grid);
+
+        assert_eq!(term.grid_to_string(), "1 2\n3456abcd\nefgh\n");
+        assert_eq!(term.grid_to_string_only_visible(), "1 2\n3456abcd\nefgh\n");
+
+        term.grid.scroll_up(&(Line(0)..Line(4)), Line(1));
+
+        assert_eq!(term.grid_to_string(), "1 2\n3456abcd\nefgh\n");
+        assert_eq!(term.grid_to_string_only_visible(), "3456abcd\nefgh\n");
     }
 
     #[test]